+++ src/settings/scrollSettings.js
@@ -43,13 +43,13 @@ export default function scrollSettings() {
     },
   ];
 
+  return settingsPage(title, items, callback);
+
   function callback(key, value) {
     appSettings.update({
       [key]: value,
     });
   }
-
-  settingsPage(title, items, callback);
 }
 
 function getScrollSpeedString(speed) {